% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MITScan_ZIP.R
\name{MITScan_ZIP}
\alias{MITScan_ZIP}
\title{Perform score-based genome-wide association analysis using zero-inflated Poisson models}
\usage{
MITScan_ZIP(Ymat, X = NULL, Z = NULL, A, perm, no_cores = 64, seed = 1)
}
\arguments{
\item{Ymat}{A matrix of OTU abundances from 16S rRNA sequencing.
Rows and columns represent OTU names and samples, respectively.}

\item{X}{A matrix of covariates related to the count component.}

\item{Z}{A matrix of covariates related to the zero component.}

\item{A}{A matrix of host gene expression from RNA-Seq.
Rows and columns represent host gene names and samples, respectively.}

\item{perm}{An integer indicating the number of permutations for genomic control.}

\item{no_cores}{An integer indicating the number of cores for parallel calculations.}

\item{seed}{An integer specifying the seed for random number generation.}
}
\value{
A matrix of score test statistics across all OTU abundances and host genes.
Rows and columns represent host gene names and OTU names, respectively.
The last row contains the median of permuted score test statistics for each OTU.
}
\description{
MITScan_ZIP() performs score-based genome-wide association analysis
across all OTU abundances and host genes using zero-inflated Poisson models,
and returns a matrix where the (i,j) element of the matrix represents
the score test statistic for the i-th host gene and the j-th OTU abundance.
}
\examples{
OTU = IBD_OTU
hostgene = IBD_hostgene
cvrt = IBD_cvrt
zip1 = MITScan_ZIP(Ymat=OTU, X=cvrt, Z=NULL, A=hostgene, perm=2, no_cores=8, seed=1)

}
